Add fallback display for multiple-level-derivable passphrases
This allows to continue managing passphrases in the UI (esp. clean up derivable passphrases), when someone screwed up the derivable passphrases tree via the API or direct entry into the database.

By now, multi-level derivation of passphrases is not usable for anything. Still, there might be usecases in the future. So, we don't restrict it yet, but instead make sure that nothing actually breaks when creating such passphrases.